+++ src/options/getMarkOptions.ts
@@ -1,7 +1,7 @@
 import { MarkOptions } from "@observablehq/plot";
-import { DuckPlot } from ".";
+import { DuckPlot } from "..";
 import { isColor } from "./getPlotOptions";
-import { borderOptions, defaultColors } from "./helpers";
+import { borderOptions, defaultColors } from "../helpers";
 
 // Get options for a specific mark (e.g., the line or area marks)
 export async function getMarkOptions(instance: DuckPlot) {
